Crate libobs_source_macro

Attribute Macros

obs_object_builder

This macro is used to generate a builder pattern for an obs source.
The attribute should be the id of the source.
The struct should have named fields, each field should have an attribute #[obs_property(type_t="your_type")].
type_t can be enum, enum_string, string, bool or int.

obs_object_impl

obs_object_updater

This macro is used to generate an updater pattern for an obs object (for example a source). For more examples look at libobs-sources